+++ glue_vispy_viewers/scatter/scatter_viewer.py
@@ -30,7 +30,11 @@ class VispyScatterViewer(BaseVispyViewer):
 
         if added:
             if first_layer_artist:
-                self.state.set_limits(*self._layer_artist_container[0].default_limits)
+                # The above call to add_data may have added subset layers, some
+                # of which may be incompatible with the data, so we need to now
+                # explicitly use the layer for the actual data object.
+                layer = self._layer_artist_container[data][0]
+                self.state.set_limits(*layer.default_limits)
                 self._ready_draw = True
 
         return added