+++ glue_vispy_viewers/volume/volume_viewer.py
@@ -175,7 +175,11 @@ class VispyVolumeViewer(BaseVispyViewer):
                 self._vispy_widget._update_limits()
 
             if first_layer_artist:
-                self.state.set_limits(*self._layer_artist_container[0].bbox)
+                # The above call to add_data may have added subset layers, some
+                # of which may be incompatible with the data, so we need to now
+                # explicitly use the layer for the actual data object.
+                layer = self._layer_artist_container[data][0]
+                self.state.set_limits(*layer.bbox)
                 self._ready_draw = True
                 self._update_slice_transform()